requirements.txt
watchdog
AfricasTalkingGateway
pyodbc
